CLN cleanup imports and shifts between the classes
